@@ -3,16 +3,18 @@
   "version": "0.0.1",
   "private": true,
   "engines": {
-    "node": "6.9.4"
+    "node": "6.10.2"
   },
   "scripts": {
-    "start": "node src/app.js",
+    "build": "tsc",
+    "start": "tsc && node dist/app.js",
     "test": "gulp test:client"
   },
   "dependencies": {
     "ajv": "4.11.6",
     "bcryptjs": "^2.4.0",
     "body-parser": "~1.15.2",
+    "caller": "^1.0.1",
     "cookie-parser": "~1.4.3",
     "core-js": "^2.4.1",
     "debug": "~2.2.0",
@@ -34,7 +36,17 @@
     "swaggerize-express": "^4.0.5"
   },
   "devDependencies": {
+    "@types/body-parser": "^1.16.4",
+    "@types/cookie-parser": "^1.3.30",
+    "@types/crypto-js": "^3.1.33",
+    "@types/express": "^4.0.36",
+    "@types/jsonwebtoken": "^7.2.1",
+    "@types/log4js": "^0.0.32",
+    "@types/mongodb": "^2.2.7",
+    "@types/mongoose": "^4.7.18",
     "@types/node": "^7.0.14",
+    "@types/swagger-parser": "^4.0.1",
+    "@types/swaggerize-express": "^4.0.28",
     "del": "^2.2.2",
     "del-cli": "^0.2.0",
     "gulp": "^3.9.1",
@@ -50,6 +62,7 @@
     "request": "^2.79.0",
     "run-sequence": "^1.2.2",
     "supertest": "^3.0.0",
+    "typescript": "^2.4.1",
     "v8-profiler": "5.7.0"
   }
 }
